+++ datacontract/imports/importer.py
@@ -44,6 +44,7 @@ class ImportFormat(str, Enum):
     redshift = "redshift"
     postgres = "postgres"
     athena = "athena"
+    s3 = "s3"
 
     @classmethod
     def get_supported_formats(cls):